+++ Runtime/TreasuredData.cs
@@ -1,4 +1,6 @@
-﻿using System.Collections.Generic;
+﻿using Newtonsoft.Json;
+using System.Collections.Generic;
+using System.Linq;
 using UnityEngine;
 
 namespace Treasured.SDK
@@ -30,9 +32,9 @@ namespace Treasured.SDK
         [SerializeField]
         private ImageQuality _quality = ImageQuality.Medium;
         [SerializeField]
-        private List<TreasuredObject> _hotspots;
+        private List<TreasuredObject> _hotspots = new List<TreasuredObject>();
         [SerializeField]
-        private List<TreasuredObject> _interactables;
+        private List<TreasuredObject> _interactables = new List<TreasuredObject>();
 
         public string Name { get => _name; set => _name = value; }
         public bool Loop { get => _loop; set => _loop = value; }
@@ -40,5 +42,16 @@ namespace Treasured.SDK
         public ImageQuality Quality { get => _quality; set => _quality = value; }
         public List<TreasuredObject> Hotspots { get => _hotspots; set => _hotspots = value; }
         public List<TreasuredObject> Interactables { get => _interactables; set => _interactables = value; }
+        /// <summary>
+        /// Returns a unique list of objects.
+        /// </summary>
+        [JsonIgnore]
+        public IEnumerable<TreasuredObject> All
+        {
+            get
+            {
+                return Hotspots.Union(Interactables);
+            }
+        }
     }
 }