+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+## Version 0.3.2 (August 17, 2021)
+- Add Visible Targets to Json
+- Add Layer Mask for export settings
 ## Version 0.3.1 (August 17, 2021)
 #### Fixes
 - Change output format jpg from jpeg and lower case the format in JSON
 - Auto-generate id for the first item in the list
+- Update image output file name
 #### Features
 - Add dropdown for selecting Target ID for `Select Object` action
 - Add button to regenerate Action ID